Implement functionality that allows to drop whole folder in ContentBrowser. Fix bug in SizeOf function that cause invalid erases.

+++ Engine/Utils/IOUtils.h
@@ -80,13 +80,13 @@ namespace MyEngine {
 
 	inline long long SizeOf(const string& value)
 	{
-		return sizeof(long long) + value.size() * sizeof(char);
+		return sizeof(int) + value.size() * sizeof(char);
 	}
 
 	template <class T>
 	inline long long SizeOf(const vector<T>& value)
 	{
-		return sizeof(long long) + value.size() * sizeof(T);
+		return sizeof(int) + value.size() * sizeof(T);
 	}
 
 #pragma endregion
